added caged landscapes
The cage is just a flat surface
surrounded by nearly vertical walls 1000 tall.

The caged mountian drops a mountain 100 tall
in the center of the cage.

Both cages go from -10,10 in x and y (plus .1 for the cage wall
thickness). The mountain covers -5,5 in both x and y and has a peak
at 0,0.